Extract resolveDefaultModel, parameterise scheme test
Address Greptile review on #2370.

- Pull `resolveDefaultModel` into a shared util so the inbox Discuss and
  Create PR hooks aren't carrying identical copies of the preview-config
  lookup. Both hooks now import from `utils/resolveDefaultModel`.
- Collapse the two prod/dev deeplink scheme tests for
  `buildCreatePrReportPrompt` into a single `it.each` block.

Skipping the analytics `created_from: "command-menu"` → `"inbox_report"`
suggestion: that label is a typed enum (`TaskCreatedFrom = "cli" |
"command-menu"`), so adding a new value would touch the analytics schema
and diverge from the Discuss flow this hook intentionally mirrors. Out of
scope for the prompt-slim PR.

Generated-By: PostHog Code
Task-Id: 4884e8d3-538d-4b19-9dc9-3b76a07df269
